Extract toggle and its mutation to a component
